requirements.txt
pinecone>=5.0.0
openai>=1.0.0
gspread>=5.10.0
oauth2client>=4.1.3
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.90.0
python-dotenv>=1.0.0
